requirements.txt
fastapi
uvicorn
langgraph
crewai
imagehash
pillow
python-dotenv
google-generativeai
chromadb
requests
